+++ packages/gits_http/CHANGELOG.md
@@ -1,3 +1,28 @@
+## 2.0.0
+
+- stable version with add cache strategy
+- add equatable for each strategy
+
+## 2.0.0-dev.3
+
+- fix fromMap for headers Map in cache wrapper
+
+## 2.0.0-dev.2
+
+- fix simply key for write to storage
+- fix cache wrapper add status_code in toMap
+
+## 2.0.0-dev.1
+
+- fix hive init
+
+## 2.0.0-dev
+
+- add cache strategy (AsyncOrCacheStrategy, CacheOrAsyncStrategy, JustAsyncStrategy, JustCacheStrategy), default JustAsyncStrategy
+- remove head method in GitsHttp
+- remove read method in GitsHttp
+- remove readBytes method in GitsHttp
+
 ## 1.1.3
 
 - update dependency gits_inspector to 2.0.1